+++ NEWS.md
@@ -1,3 +1,13 @@
+Version 0.2.14, 2015-10-30
+-----------------------------------------------------------------------------------
+
+FEATURES / CHANGES
+
+- Addition of `parLapplyW()` and `pddply()`
+- Various edits to prepare `Smisc` for submission to CRAN
+- Support for Cairo graphics for `openDevice()` is removed
+
+
 Version 0.2.13, 2015-10-12
 -----------------------------------------------------------------------------------